pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stta"
version = "0.1.0"
description = "STT Analytics Platform - Speech-to-Text transcript analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "STT Analytics Team"}
]
keywords = ["speech-to-text", "analytics", "call-center", "transcription"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pandas>=2.2.0,<3.0.0",
    "pyarrow>=17.0.0,<19.0.0",
    "numpy>=2.0.0,<3.0.0",
    "pandera>=0.18.0,<0.19.0",
    "pydantic>=2.7.0,<3.0.0",
    "plotly>=5.24.0,<6.0.0",
    "kaleido>=0.2.1,<0.3.0",
    "streamlit>=1.39.0,<2.0.0",
    "reportlab>=4.0.0,<5.0.0",
    "typer>=0.12.0,<0.13.0",
    "rich>=13.7.0,<14.0.0",
    "loguru>=0.7.0,<0.8.0",
    "orjson>=3.10.0,<4.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "python-dateutil>=2.9.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "hypothesis>=6.112.0,<7.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.5.0,<0.6.0",
    "mypy>=1.10.0,<2.0.0",
]

[project.scripts]
stta = "stta.cli:app"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--verbose --strict-markers"
